move assets folder and fix lint errors
