+++ packages/react/ds/src/assets/logos/logo-small.tsx
