+++ packages/react/ds/src/assets/logos/logo.tsx
